pyproject.toml
[tool.poetry]
name = "solid-backend"
version = "0.3.0"

description = "Clean Django app for e-learning application with..."
authors = [
	"Christian Grossmüller <chgad.games@gmail.com>",
	"Timotheus Kozlowski <kozlowski@itp.uni-frankfurt.de>"
]
license = "MIT"

readme = "README.md"
homepage = "https://github.com/zentrumnawi/solid-backend"
repository = "https://github.com/zentrumnawi/solid-backend"


[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
djangorestframework = "3.11.0"
django-mptt = "0.11.0"
pillow = "7.1.2"
django-cleanup = "4.0.0"
psycopg2-binary = "^2.8.5"
mutagen = "^1.44.0"
django-stdimage = "^5.1.1"
django-taggit = "^1.5.1"
django-filter = "^21.1"


[tool.poetry.dev-dependencies]
gunicorn = "20.0.4"
django-environ = "0.4.5"
werkzeug = "1.0.1"
django-debug-toolbar = "2.2"
ipdb = "0.13.2"
pytest = "5.3.5"
pytest-django = "3.9.0"
django-extensions = "2.2.9"
pytest-isort = "1.0.0"
pytest-sugar = "0.9.3"
coverage = "5.1"
django-coverage-plugin = "1.8.0"
pytest-cov = "2.9.0"
codecov = "2.1.4"
black = "19.10b0"
flake8 = "*"
drf-yasg = "^1.20.0"


[tool.poetry.group.extras.dependencies]
drf-spectacular = "^0.25.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
